Revert "Revert "Minor change to function""
This reverts commit 28429072b767c752de67cce7c27c45d218057701.

+++ src/java/com/messagereplicator/util/Function.java
@@ -15,6 +15,6 @@ public abstract class Function {
 	 * @param x
 	 * @return f(x)
 	 */
-	public abstract double perform(float x);
+	public abstract float perform(float x);
 	
 }
